requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
moviepy==1.0.3
edge-tts==6.1.12
google-generativeai==0.8.3
google-api-python-client==2.150.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
ffmpeg-python==0.2.0
pydub==0.25.1
pillow==11.0.0
numpy==1.26.4
requests==2.32.3
nltk==3.9.1
spacy==3.8.2
serpapi==0.1.5
groq==0.11.0
proglog==0.1.10
faster-whisper==1.0.3
imageio-ffmpeg==0.5.1
python-dotenv==1.0.1
pyjwt==2.9.0
